task8_chatbot/requirements.txt
# Task 8 - Chatbot
python-dotenv>=1.0.0
together>=0.2.0
requests>=2.28.0
